requirements.txt
aiogram==3.13.1
fake_useragent==1.5.1
google_api_python_client==2.149.0
playwright==1.47.0
protobuf==5.28.3
python-dotenv==1.0.1
